Create: Main Code
Create: Main Batch Code has been added to the Repository.

It still needs some improvement. It works and the user can interact with the menu. But the user has to be able to go back to the previous menu when a command has been ran.

Instead of closing and re-opening.