Tweak: Visual Appearance - Code
Tweak: Updated the visual appearance of the code. So it's prettier and easier to read for the user what he/she is about to do.

The code was a bit hard to read in the past, we hope to have adressed that issue.

There might be a possibility that there will be more reworks on the visual front.