Tweak: Trace Log Function
Tweak: A tweak has been made to the useability of the function.

Thanks to @Stensel8 the error in the code had been found. Thanks for adressing this issue and your further help.

Besides that the code has had some small QoL improvements regarding to the layout of the user 'UI.'