Update: Documentation
Update: Updated the documentation about the new project: pcHealthPlus - official repository name is pcHealthPlus-VS.

In the main menu there is a new option, this options brings you to the README.md of the 'new' version of pcHealth. This new version of pcHealth is still in development and not even near finishing, but for those who care and those who enjoy using pcHealth.. they will know with this function about the development of that.