Tweak: PowerToys
Tweak: Made a tweak in the new added function; PowerToys. When the user tried to install this feature through the script, it would break. Since the script was trying to run a PowerShell command in CMD.

This issue has been resolved. The script now emulates the command in CMD as a PowerShell command. Sounds complicated but this is a standard issue which I should have implemented from the start. But lets say I made a minor error by rushing it.

Only possible error that might occur is that it could give a 'winget' error. It occured to me on one of my test VM's. From the 3, 1 had a failure.