Add: Preparation of Bootrec Script
Add: Added a new .PS1 script that starts the boot recovery process. If users are stuck at the Windows Blue Screen (for as long as it is Blue) they can kick-off the script from the USB/CMD line.

It is written in .PS1 for conveniance but the commands are originally from Batch.

Ingestion of the Powershell module needs to be added in W10 to work with this feature. Besides that, the function/script does work as intended but it needs to be given some love.